Raphael

+++ src/Routes/PizzaRoute.js
@@ -1,5 +1,7 @@
 import express from 'express';
 import PizzaController from '../Controllers/PizzaController';
+import { error } from '../returnFunc';
+import JWT from 'jsonwebtoken';
 
 export const anonymeRoutePizza = express.Router();
 export const adminRoutePizza = express.Router();
@@ -19,31 +21,68 @@ anonymeRoutePizza.route('/category/:cat')
         const Pizza = await PizzaController.getPizzaByCat(req.params.cat);
         res.json(Pizza);
     })
-/*
-anonymeRoutePizza.route('/add')
+
+adminRoutePizza.route('/add')
     .post(async (req, res) => {
+        if(!req.body.categoryId && !req.body.sizeId) res.json(error('Error, empty value categoryId or sizeId'));
+
+        const decode = await JWT.decode(req.headers['x-access-token'], {complete: true});
+        const { role } = decode.payload;
+        
         const pizza = {
-            price: req.params.price,
-            size: req.params.size,
-            composition: {
-                sauce : req.params.sauce,
-                cheese: false 
-            },
-            ingredient: {
-                connect: {
-                    id: 'ck79bqj5g001j0739verzunk9'
+            name: req.body.name,
+            composition: req.body.composition,
+            size: {
+                connect : {
+                   id: req.body.sizeId
                 }
             },
             category: {
                 connect: {
-                    id: 'ck79bqmjd001t07396lyh6cj0'
+                    id: req.body.categoryId
                 }
             }
+        };
+
+        if (role.indexOf('ROLE_ADMIN') !== -1){
+            const addPizza = await PizzaController.createPizza(pizza);
+            res.json(addPizza);
+        }else {
+            res.json(error('You are not an admin'));
         }
-    })*/
+        
+    })
+
 adminRoutePizza.route('/delete/:id')
     .delete(async (req, res) => {
-        const pizzaParam = {id : req.params.id};
-        const deletePizza = await PizzaController.deletePizzaById(pizzaParam);
-        res.json(deletePizza) ;
+        const decode = await JWT.decode(req.headers['x-access-token'], {complete: true});
+        const { role } = decode.payload;
+
+        if (role.indexOf('ROLE_ADMIN') !== -1){
+            const pizzaParam = {id : req.params.id};
+            const deletePizza = await PizzaController.deletePizzaById(pizzaParam);
+            res.json(deletePizza);
+        }else {
+            res.json(error('You are not an admin'));
+        }
+    })
+
+adminRoutePizza.route('/update/:id')
+    .put(async (req, res) => {
+        const decode = await JWT.decode(req.headers['x-access-token'], {complete: true});
+        const { role } = decode.payload;
+
+        if (role.indexOf('ROLE_ADMIN') !== -1){
+            const updateObject = {
+                id: req.params.id,
+                composition: req.body.composition,
+                sizeId : req.body.sizeId,
+                categoryId: req.body.categoryId
+            }
+            
+            const updatePizza = await PizzaController.updatePizzaById(updateObject);
+            res.json(updatePizza);
+        }else {
+            res.json(error('You are not an admin'));
+        }
     })
